requirements.txt
torch
torchvision
detectron2
fastapi
uvicorn
google-cloud-bigquery
pillow
requests